requirements.txt
streamlit
vosk
sounddevice
numpy
scipy
groq
python-dotenv
fpdf
requests
tqdm
audio-recorder-streamlit